Fetch service rule items with service rule.
Adds a bi-direction mapping to model and sub-select fetching to DAO.

Also changes item mappings to be BigDecimal at persistence layer (which
they already were at the API model layer).

+++ backend/src/main/java/org/ccts/balancingact/model/db/ServiceTaskItemEntity.java
@@ -30,11 +30,11 @@ public class ServiceTaskItemEntity extends BaseEntity {
         return item.getDescription();
     }
 
-    public int getQuantity() {
+    public BigDecimal getQuantity() {
         return item.getQuantity();
     }
 
-    public float getRate() {
+    public BigDecimal getRate() {
         return item.getRate();
     }
